refactor: replace old modules with ones based on version_set
