+++ src/internal/mod.rs
@@ -4,10 +4,7 @@
 
 pub mod arena;
 pub mod core;
-pub mod core_bis;
 pub mod incompatibility;
-pub mod incompatibility_bis;
 pub mod partial_solution;
-pub mod partial_solution_bis;
 pub mod small_map;
 pub mod small_vec;
