+++ src/term.rs
@@ -3,38 +3,37 @@
 //! A term is the fundamental unit of operation of the PubGrub algorithm.
 //! It is a positive or negative expression regarding a set of versions.
 
-use crate::range::Range;
-use crate::version::Version;
-use std::fmt;
+use crate::version_set::VersionSet;
+use std::fmt::{self, Display};
 
 ///  A positive or negative expression regarding a set of versions.
 #[derive(Debug, Clone, Eq, PartialEq)]
-pub enum Term<V: Version> {
+pub enum Term<VS: VersionSet> {
     /// For example, "1.0.0 <= v < 2.0.0" is a positive expression
     /// that is evaluated true if a version is selected
     /// and comprised between version 1.0.0 and version 2.0.0.
-    Positive(Range<V>),
+    Positive(VS),
     /// The term "not v < 3.0.0" is a negative expression
     /// that is evaluated true if a version is selected >= 3.0.0
     /// or if no version is selected at all.
-    Negative(Range<V>),
+    Negative(VS),
 }
 
 /// Base methods.
-impl<V: Version> Term<V> {
+impl<VS: VersionSet> Term<VS> {
     /// A term that is always true.
     pub(crate) fn any() -> Self {
-        Self::Negative(Range::none())
+        Self::Negative(VS::empty())
     }
 
     /// A term that is never true.
     pub(crate) fn empty() -> Self {
-        Self::Positive(Range::none())
+        Self::Positive(VS::empty())
     }
 
     /// A positive term containing exactly that version.
-    pub(crate) fn exact(version: V) -> Self {
-        Self::Positive(Range::exact(version))
+    pub(crate) fn exact(version: VS::V) -> Self {
+        Self::Positive(VS::singleton(version))
     }
 
     /// Simply check if a term is positive.
@@ -50,41 +49,41 @@ impl<V: Version> Term<V> {
     /// the opposite of the evaluation of the original one.
     pub(crate) fn negate(&self) -> Self {
         match self {
-            Self::Positive(range) => Self::Negative(range.clone()),
-            Self::Negative(range) => Self::Positive(range.clone()),
+            Self::Positive(set) => Self::Negative(set.clone()),
+            Self::Negative(set) => Self::Positive(set.clone()),
         }
     }
 
     /// Evaluate a term regarding a given choice of version.
-    pub(crate) fn contains(&self, v: &V) -> bool {
+    pub(crate) fn contains(&self, v: &VS::V) -> bool {
         match self {
-            Self::Positive(range) => range.contains(v),
-            Self::Negative(range) => !(range.contains(v)),
+            Self::Positive(set) => set.contains(v),
+            Self::Negative(set) => !(set.contains(v)),
         }
     }
 
-    /// Unwrap the range contains in a positive term.
-    /// Will panic if used on a negative range.
-    pub(crate) fn unwrap_positive(&self) -> &Range<V> {
+    /// Unwrap the set contained in a positive term.
+    /// Will panic if used on a negative set.
+    pub(crate) fn unwrap_positive(&self) -> &VS {
         match self {
-            Self::Positive(range) => range,
-            _ => panic!("Negative term cannot unwrap positive range"),
+            Self::Positive(set) => set,
+            _ => panic!("Negative term cannot unwrap positive set"),
         }
     }
 }
 
 /// Set operations with terms.
-impl<V: Version> Term<V> {
+impl<VS: VersionSet> Term<VS> {
     /// Compute the intersection of two terms.
     /// If at least one term is positive, the intersection is also positive.
-    pub(crate) fn intersection(&self, other: &Term<V>) -> Term<V> {
+    pub(crate) fn intersection(&self, other: &Self) -> Self {
         match (self, other) {
             (Self::Positive(r1), Self::Positive(r2)) => Self::Positive(r1.intersection(r2)),
             (Self::Positive(r1), Self::Negative(r2)) => {
-                Self::Positive(r1.intersection(&r2.negate()))
+                Self::Positive(r1.intersection(&r2.complement()))
             }
             (Self::Negative(r1), Self::Positive(r2)) => {
-                Self::Positive(r1.negate().intersection(r2))
+                Self::Positive(r1.complement().intersection(r2))
             }
             (Self::Negative(r1), Self::Negative(r2)) => Self::Negative(r1.union(r2)),
         }
@@ -92,14 +91,14 @@ impl<V: Version> Term<V> {
 
     /// Compute the union of two terms.
     /// If at least one term is negative, the union is also negative.
-    pub(crate) fn union(&self, other: &Term<V>) -> Term<V> {
+    pub(crate) fn union(&self, other: &Self) -> Self {
         (self.negate().intersection(&other.negate())).negate()
     }
 
     /// Indicate if this term is a subset of another term.
     /// Just like for sets, we say that t1 is a subset of t2
     /// if and only if t1 ∩ t2 = t1.
-    pub(crate) fn subset_of(&self, other: &Term<V>) -> bool {
+    pub(crate) fn subset_of(&self, other: &Self) -> bool {
         self == &self.intersection(other)
     }
 }
@@ -120,7 +119,7 @@ pub(crate) enum Relation {
 }
 
 /// Relation between terms.
-impl<'a, V: 'a + Version> Term<V> {
+impl<VS: VersionSet> Term<VS> {
     /// Check if a set of terms satisfies this term.
     ///
     /// We say that a set of terms S "satisfies" a term t
@@ -129,7 +128,7 @@ impl<'a, V: 'a + Version> Term<V> {
     /// It turns out that this can also be expressed with set operations:
     ///    S satisfies t if and only if  ⋂ S ⊆ t
     #[cfg(test)]
-    fn satisfied_by(&self, terms_intersection: &Term<V>) -> bool {
+    fn satisfied_by(&self, terms_intersection: &Self) -> bool {
         terms_intersection.subset_of(self)
     }
 
@@ -142,13 +141,13 @@ impl<'a, V: 'a + Version> Term<V> {
     ///    S contradicts t if and only if ⋂ S is disjoint with t
     ///    S contradicts t if and only if  (⋂ S) ⋂ t = ∅
     #[cfg(test)]
-    fn contradicted_by(&self, terms_intersection: &Term<V>) -> bool {
+    fn contradicted_by(&self, terms_intersection: &Self) -> bool {
         terms_intersection.intersection(self) == Self::empty()
     }
 
     /// Check if a set of terms satisfies or contradicts a given term.
     /// Otherwise the relation is inconclusive.
-    pub(crate) fn relation_with(&self, other_terms_intersection: &Term<V>) -> Relation {
+    pub(crate) fn relation_with(&self, other_terms_intersection: &Self) -> Relation {
         let full_intersection = self.intersection(other_terms_intersection);
         if &full_intersection == other_terms_intersection {
             Relation::Satisfied
@@ -160,19 +159,20 @@ impl<'a, V: 'a + Version> Term<V> {
     }
 }
 
-impl<V: Version> AsRef<Term<V>> for Term<V> {
-    fn as_ref(&self) -> &Term<V> {
-        &self
+// impl<V: Version> AsRef<Term<V>> for Term<V> {
+impl<VS: VersionSet> AsRef<Self> for Term<VS> {
+    fn as_ref(&self) -> &Self {
+        self
     }
 }
 
 // REPORT ######################################################################
 
-impl<V: Version + fmt::Display> fmt::Display for Term<V> {
+impl<VS: VersionSet + Display> Display for Term<VS> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
-            Self::Positive(range) => write!(f, "{}", range),
-            Self::Negative(range) => write!(f, "Not ( {} )", range),
+            Self::Positive(set) => write!(f, "{}", set),
+            Self::Negative(set) => write!(f, "Not ( {} )", set),
         }
     }
 }
@@ -182,10 +182,11 @@ impl<V: Version + fmt::Display> fmt::Display for Term<V> {
 #[cfg(test)]
 pub mod tests {
     use super::*;
+    use crate::range::Range;
     use crate::version::NumberVersion;
     use proptest::prelude::*;
 
-    pub fn strategy() -> impl Strategy<Value = Term<NumberVersion>> {
+    pub fn strategy() -> impl Strategy<Value = Term<Range<NumberVersion>>> {
         prop_oneof![
             crate::range::tests::strategy().prop_map(Term::Positive),
             crate::range::tests::strategy().prop_map(Term::Negative),
